rename predicate
