+++ src/Query/Predicate/Column.php
@@ -1,10 +1,10 @@
 <?php
 
-namespace Bank\Query\Clause;
+namespace Bank\Query\Predicate;
 
 /**
  * Class Column
- * @package Bank\Query\Clause
+ * @package Bank\Query\Predicate
  */
 class Column
 {
